chore!: upgraded to Spring Boot 4
BREAKING CHANGE: dropping support for Spring Boot 3 and Jackson 2

+++ framework-spring-boot-autoconfigure/src/main/java/com/opencqrs/framework/eventhandler/EventHandlingProperties.java
@@ -179,7 +179,7 @@ public record EventHandlingProperties(ProcessorSettings standard, Map<String, Pr
                 Duration maxInterval,
                 Duration maxElapsedTime,
                 Double multiplier,
-                Integer maxAttempts) {
+                Long maxAttempts) {
 
             /** The back-off policy. */
             public enum Policy {
